airflow/requirements.txt
apache-airflow-providers-apache-kafka==2.7.0
apache-airflow-providers-apache-spark==4.7.1
cassandra-driver==3.29.0
